pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-benchmark-dashboard"
version = "0.1.0"
description = "Data-quality-first dashboard for AI benchmark progress tracking"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "AI Benchmark Dashboard Team" }
]

dependencies = [
    # Data processing
    "polars>=1.0.0",
    "duckdb>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",

    # Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # API
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",

    # Dashboard
    "streamlit>=1.30.0",
    "plotly>=5.18.0",

    # HTTP/Scraping
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",

    # Projections
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",

    # Utilities
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

[project.scripts]
benchmark-dashboard = "src.dashboard.app:main"
benchmark-update = "scripts.update_data:main"
benchmark-api = "src.api.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
